Create BuySellStock.py
We are given an array prices where prices[i] is the price of a given stock on the i th day.

We want to maximize your profit by choosing a single day to buy one stock and choosing a different day in the future to sell that stock.

Return the maximum profit you can achieve from this transaction. If you cannot achieve any profit, return 0.